Add interoperability macro for NSRectAlignment

+++ IBPCollectionViewCompositionalLayout/IBPNSCollectionLayoutAnchor.m
@@ -112,7 +112,7 @@ IBPNSDirectionalRectEdge EdgesFromAnchorPoint(CGPoint anchorPoint);
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"<NSCollectionLayoutAnchor %p: edges=%lu; offset=%@; anchorPoint=%@>", self, self.edges, NSStringFromCGPoint(self.offset), NSStringFromCGPoint(self.offset)];
+    return [NSString stringWithFormat:@"<NSCollectionLayoutAnchor %p: edges=%lu; offset=%@; anchorPoint=%@>", self, (unsigned long)self.edges, NSStringFromCGPoint(self.offset), NSStringFromCGPoint(self.offset)];
 }
 
 @end
